+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 功能构建
 
+- `dwarfeng-dct-api` 子模块类优化注释、文档注释格式、代码换行格式。
+  - com.dwarfeng.dct.api.integration.configuration.ServiceExceptionMapperConfiguration。
+
 - `dwarfeng-dct-core` 子模块类优化注释、文档注释格式、代码换行格式。
   - com.dwarfeng.dct.node.configuration.DataCodingConfigDefinitionParser。
   - com.dwarfeng.dct.node.configuration.DataCodingQosDefinitionParser。